+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Change Log
 ## v0.2.3
 
 - updated pacakging and build system resolve install issues
-- 
+- show usage and exit if no search query provided
 
 ## v0.2.2